# Bulan 1-2/Minggu 5-6/Project/inventory-stock-management-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Stock Management API",
        "description": "REST API untuk mengelola data inventaris, produk, supplier, warehouse, dan stok.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/warehouses": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get All Warehouses",
                "operationId": "0c48cbbd5731b5e56eda7c180576670b",
                "responses": {
                    "200": {
                        "description": "List of warehouses"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Create Warehouse",
                "operationId": "c1b0005f758b5892517cade2be77a1d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Warehouse A"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Subang"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Main Warehouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Warehouse created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/warehouses/{id}": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get Warehouse Detail",
                "operationId": "9f201c10112ab3904ca3aa4e71e63e89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse detail"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Update Warehouse",
                "operationId": "8d7614858df133388092bda00a4be309",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Warehouse B"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated Warehouse"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Warehouse updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Delete Warehouse",
                "operationId": "45d106d3e2367cdc02fc89c04fc702c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}